requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
openai>=1.35.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── PDF Parsing ───────────────────────────────────────────────────────────────
pdfplumber>=0.11.0
pypdf>=4.0.0

# ── PDF Generation ────────────────────────────────────────────────────────────
reportlab>=4.2.0
